+++ crates/pcode-ops/src/validate.rs
